+++ spessasynth_core/src/synthesizer/audio_engine/channel/portamento_time.c
@@ -17,29 +17,29 @@
 
 typedef struct SS_PortamentoLookup {
 	uint8_t key;
-	float value;
+	double value;
 } SS_PortamentoLookup;
 
 static const SS_PortamentoLookup portamento_lookup_table[] = {
 	{ 0, 0.0 },
-	{ 1, 0.006f },
-	{ 2, 0.023f },
-	{ 4, 0.05f },
-	{ 8, 0.11f },
-	{ 16, 0.25f },
-	{ 32, 0.5f },
-	{ 64, 2.06f },
-	{ 80, 4.2f },
-	{ 96, 8.4f },
-	{ 112, 19.5f },
-	{ 116, 26.7f },
-	{ 120, 40.0f },
-	{ 124, 80.0f },
-	{ 127, 480.0f }
+	{ 1, 0.006 },
+	{ 2, 0.023 },
+	{ 4, 0.05 },
+	{ 8, 0.11 },
+	{ 16, 0.25 },
+	{ 32, 0.5 },
+	{ 64, 2.06 },
+	{ 80, 4.2 },
+	{ 96, 8.4 },
+	{ 112, 19.5 },
+	{ 116, 26.7 },
+	{ 120, 40.0 },
+	{ 124, 80.0 },
+	{ 127, 480.0 }
 };
 static const int portamento_lookup_table_count = sizeof(portamento_lookup_table) / sizeof(portamento_lookup_table[0]);
 
-static float ss_portamento_get_lookup(int time) {
+static double ss_portamento_get_lookup(int time) {
 	const int count = portamento_lookup_table_count;
 	for(int i = 0; i < count; i++) {
 		if(portamento_lookup_table[i].key == time)
@@ -64,15 +64,15 @@ static float ss_portamento_get_lookup(int time) {
 	}
 
 	if(lower != -1 && upper != -1) {
-		const float lowerTime = portamento_lookup_table[lower_index].value;
-		const float upperTime = portamento_lookup_table[upper_index].value;
+		const double lowerTime = portamento_lookup_table[lower_index].value;
+		const double upperTime = portamento_lookup_table[upper_index].value;
 
-		return (lowerTime + ((float)(time - lower) * (upperTime - lowerTime)) / (float)(upper - lower));
+		return (lowerTime + ((double)(time - lower) * (upperTime - lowerTime)) / (double)(upper - lower));
 	}
 
 	return 0;
 }
 
-float ss_portamento_time_to_seconds(float portamento_time, float distance) {
-	return ss_portamento_get_lookup((int)portamento_time) * (distance / 36.0f);
+double ss_portamento_time_to_seconds(double portamento_time, double distance) {
+	return ss_portamento_get_lookup((int)portamento_time) * (distance / 36.0);
 }
